Fix sort for inputs with very big headers
These can't be written as BAM files, but can using SAM or CRAMv4.
Extend the check on output compatibility to check for files
with very big headers, and use SAM for temporary files in this
case.

The temporary file fix comes with the disadvantage that the
SAM reader hangs on to a copy of the header, resulting in
the need to store multiple copies when merging the temporary
files.  This will be fixed in a future update that removes the
need to read and write the headers to the temp. files.